+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -4255,7 +4255,7 @@ export function schemaFromJSON(
 /** @internal */
 export const Tools$inboundSchema: z.ZodType<Tools, z.ZodTypeDef, unknown> = z
   .object({
-    id: z.string().default("01KAXXWAVCJE3AZ7BYS51PDTJY"),
+    id: z.string().default("01KAZCQ509X8YP4B90M5MVYKBW"),
     name: z.string(),
     description: z.string().optional(),
     schema: z.lazy(() => Schema$inboundSchema),
@@ -4274,7 +4274,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KAXXWAVCJE3AZ7BYS51PDTJY"),
+  id: z.string().default("01KAZCQ509X8YP4B90M5MVYKBW"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => Schema$outboundSchema),