+++ packages/orq-rc/RELEASES.md
@@ -5399,3 +5399,13 @@ Based on:
 - [typescript v4.1.0-rc.5] packages/orq-rc
 ### Releases
 - [NPM v4.1.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.5 - packages/orq-rc
+
+## 2025-11-26 06:15:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.661.0 (2.762.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.1.0-rc.7] packages/orq-rc
+### Releases
+- [NPM v4.1.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.7 - packages/orq-rc